requirements.txt
anthropic==0.40.0
requests==2.32.3
feedparser==6.0.11
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
resend==2.10.0
